Design: 채팅방 가격변경 바텀시트 제작 및 기본 동작 추가
