+++ src/types/priceTable.ts
@@ -1,0 +1,9 @@
+export interface productData {
+  code: string;
+  image: string;
+  accommodationName: string;
+  roomName: string;
+  checkInDate: string;
+  checkOutDate: string;
+  policyNumber: number;
+}
